requirements.txt
python_dotenv==1.0.1
requests==2.32.3
djangorestframework==3.15.2
Django==5.1.3